+++ crates/puffin-fs/Cargo.toml
@@ -13,7 +13,6 @@ license = { workspace = true }
 workspace = true
 
 [dependencies]
-puffin-path = { path = "../puffin-path" }
 puffin-warnings = { path = "../puffin-warnings" }
 
 dunce = { workspace = true }
