+++ crates/puffin-fs/src/path.rs
