+++ crates/puffin-git/src/git.rs
@@ -9,7 +9,7 @@ use std::{env, str};
 use anyhow::{anyhow, Context as _, Result};
 use cargo_util::{paths, ProcessBuilder};
 use git2::{self, ErrorClass, ObjectType};
-use puffin_path::NormalizedDisplay;
+use puffin_fs::NormalizedDisplay;
 use reqwest::Client;
 use reqwest::StatusCode;
 use tracing::{debug, warn};
