--- crates/puffin-path/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "puffin-path"
-version = "0.0.1"
-edition = "2021"
-description = "Representations for filesystem paths."
-
-[dependencies]
-dunce = { workspace = true }
