+++ crates/puffin/src/commands/freeze.rs
@@ -8,9 +8,9 @@ use tracing::debug;
 use distribution_types::Name;
 use platform_host::Platform;
 use puffin_cache::Cache;
+use puffin_fs::NormalizedDisplay;
 use puffin_installer::SitePackages;
 use puffin_interpreter::Virtualenv;
-use puffin_path::NormalizedDisplay;
 
 use crate::commands::ExitStatus;
 use crate::printer::Printer;
