+++ crates/requirements-txt/Cargo.toml
@@ -15,8 +15,8 @@ workspace = true
 [dependencies]
 pep440_rs = { path = "../pep440-rs", features = ["serde"] }
 pep508_rs = { path = "../pep508-rs", features = ["serde"] }
+puffin-fs = { path = "../puffin-fs" }
 puffin-normalize = { path = "../puffin-normalize" }
-puffin-path = { path = "../puffin-path" }
 
 fs-err = { workspace = true }
 once_cell = { workspace = true }
